% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{formula_hlm_to_lmer}
\alias{formula_hlm_to_lmer}
\title{HLM formula to lmer formula}
\usage{
formula_hlm_to_lmer(formula_char, correlated_ranef = T)
}
\arguments{
\item{formula_char}{.}

\item{correlated_ranef}{.}
}
\description{
Converts a character vector representing a Multilevel/Hierarchical Linear Model formula to a mixed model formula for using with lmer (package lme4)
}
