% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_hlm.R
\name{summary_hlm}
\alias{summary_hlm}
\title{Tidy HLM Summary for a lmer object}
\usage{
summary_hlm(model)
}
\arguments{
\item{model}{A model estimated with lmer (package lme4), using a HLM formula converted using formula_hlm_to_lmer.}
}
\description{
Tidy HLM Summary for a lmer object
}
