% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_hlm.R
\name{tidy_hlm}
\alias{tidy_hlm}
\title{Extracts a Tidy list of summary objects from a lmer object}
\usage{
tidy_hlm(model)
}
\arguments{
\item{model}{A model estimated with lmer (package lme4), using a HLM formula converted using formula_hlm_to_lmer.}
}
\description{
Extracts and organizes the results from a lmer model into a HLM summary
}
